refactor: move scroller to widgets
