+++ selfdrive/ui/layouts/settings/toggles.py
@@ -1,7 +1,7 @@
 from openpilot.common.params import Params
-from openpilot.system.ui.lib.scroller import Scroller
 from openpilot.system.ui.lib.widget import Widget
 from openpilot.system.ui.widgets.list_view import multiple_button_item, toggle_item
+from openpilot.system.ui.widgets.scroller import Scroller
 
 # Description constants
 DESCRIPTIONS = {
